Rewrote sequential A-Star.

+++ src/Heuristic.c
@@ -23,7 +23,7 @@ static double Harversine(Coord coord1, Coord coord2){
     return R * d * 1000; // meters
 }
 
-double Hcoord(Coord coord1, Coord coord2){
+int Hcoord(Coord coord1, Coord coord2){
     return sqrt(pow(coord1->c1 - coord2->c1,2) + pow(coord1->c2 - coord2->c2,2));
     //return Harversine(coord1, coord2);
 }
